chore: rename new config file to make a nicer diff
